Former-commit-id: 59fb268329deb25a8eafeaa9db30238369380de7 [formerly 61113eecc1d453c79d25a6e9563c2fd7a0b667ee]
Former-commit-id: 50eea99390ed0e37fd01a57c1207ea837b618c79

+++ app/Http/Controllers/ContractController.php
@@ -370,7 +370,10 @@ class ContractController extends Controller
     public function removefile(Request $request, Contract $contract)
     {
         $media = $contract->getMedia('document')->where('id', $request->input('id'))->first();
-        $media->delete();
+        if(!empty($media) == 0){
+            $media->delete();
+        }
+        
 
         return response()->json(null, 204);
     }
